backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
bcrypt==4.1.2

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# RAG & AI
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19
chromadb==0.4.22
tiktoken==0.5.2
sentence-transformers==2.3.1

# Document Processing
pypdf==4.0.1
pdfplumber==0.10.3

# Security & Utilities
python-dotenv==1.0.1
httpx==0.26.0
tenacity==8.2.3

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
